GH2883 move basicInfoService in order not to lose code coverage
